feat: allow to unlock at same time
